+++ compiler/primative/main.js
@@ -5,8 +5,10 @@ const Function = require('./../component/function.js');
 
 const Array_Template = require('./array.js');
 const Blank = require('./blank.js');
+const Either = require('./either.js');
 const Static_Cast = require('./static_cast.js');
 const Bitcast = require('./bitcast.js');
+const Printf = require('./printf.js');
 const SizeOf = require('./sizeof.js');
 const types = require('./types.js');
 
@@ -28,8 +30,8 @@ function Generate (ctx) {
 	file.names.bitcast = new Bitcast(file);
 	file.names.sizeof = new SizeOf(file);
 	file.names.Blank = new Blank(file);
-
-	// file.names.Array = new Array_Template(file);
+	file.names.Either = new Either(file);
+	file.names.printf = new Printf(file);
 
 	ctx.inject(file);
 }